refactor(hooks): extract top selling products logic into custom hook
